CONTRIBUTED: electrum-ftc.desktop: improve user installation
Before, the desktop entry file would only work correctly
if the user has added `$HOME/.local/bin` to his or her
PATH. The PATH variable is an implementation detail and
the user shouldn't be exposed to it.

Now, the shell is used to look for the aforementioned
folder. This also works for system-wide installations.